smol restructuring
